configs moved to shared modules directory
