--- src/app/core/config/consts/index.ts
@@ -1,2 +1,0 @@
-export * from './config';
-export * from './css-classes';
