+++ src/app/shared/config/consts/config.ts
@@ -1,0 +1,22 @@
+
+export enum CONSTS_VARS {
+  cookieXSRFTokenKey = 'XSRF-TOKEN=',
+  applicationDocumentTitle = 'Wasp Logística',
+  forbiddenUsername = 'administrator'
+}
+
+export enum PEDIDOS_TYPE_NAMES {
+  dar = 'D.A.R',
+  imposicao = 'Imposição'
+}
+
+export enum MEDIA_BREAKPOINTS {
+  small = 'all and (max-width: 768px)',
+  medium = 'all and (max-width: 1000px)',
+  large = 'all and (max-width: 1300px)'
+}
+
+export enum LAYOUT_FLAGS {
+  headerHeight = '128px',
+  footerHeight = '130px'
+}
