+++ src/app/shared/config/consts/index.ts
@@ -1,0 +1,4 @@
+export * from './config';
+export * from './css';
+export * from './storage';
+export * from './types';
