+++ src/app/shared/config/consts/storage.ts
@@ -1,0 +1,10 @@
+
+export enum LOCAL_STORAGE {
+  themeKey = 'user-theme',
+  sessionKey = 'token',
+  sessionTTL = 'expirationTime',
+  sessionUsername = 'username',
+  sessionEmail = 'email',
+  sessionAccessType = 'accessType',
+  currentUserView = 'currentView'
+}
