+++ src/app/shared/config/index.ts
@@ -1,0 +1,2 @@
+export * from './consts';
+export * from './routes';
