+++ src/app/shared/config/routes/app-routes.ts
